Award 7 devotion points at the start of the game...
Currently, devotions are only rewarded hours into a fresh run (on reaching mile-stones or through an expensive purchase).
To speed things up, allow players to start with their first constellation unlocked.
This is especially important for early devotion skills which will start leveling up with the player (very quickly early on).
The previous balance was fine when aiming at around 4 hours for good progress on a run but now I'm leaning more towards 2 as a general guide-line.